d/postinst: run reinit unconditionally upon configure
triggers are not accumulated (and `postinst triggered` is not called)
if the package is itself upgraded with the same apt/dpkg invocation.
(i.e. `apt install --reinstall systemd-boot-efi` triggers while
`apt install --reinstall systemd-boot-efi proxmox-kernel-helper` does
not).

This was noticed while upgrading from bookworm to trixie (the reinit
was not run, and the ESPs contained the bootloaders from bookworm).

The behavior is documented in: /usr/share/doc/dpkg/spec/triggers.txt
```
Or to put it another way, triggered actions are considered irrelevant
if the interested package <I> is not configured. When <I>'s postinst is
called with ‘configure’, it must do whatever actions are necessary to
deal with any trigger activations which might have occurred while it
was not configured, just as if the package was being configured for
the first time.
```

Reported-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Suggested-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
(cherry picked from commit 8c9cb994439d70cd781e9f7af89d9a9444dcdb98)
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>